// src/app/Search/type.ts
"use client"

export interface Criteria {
  selectedCity: string;
  selectedDistrict: string;
  selectedInitialCount: string;
  selectedMaxCount: string;
  selectedInitialArea: string;
  selectedMaxArea: string;
  selectedLayout: string[];
  selectedPropertyTypes: string[];
  selectedOthers: string[];
  selectedHouseTypes: string[];
}